Crate gfx_hal

Low-level graphics abstraction for Rust. Mostly operates on data, not types. Designed for use by libraries and higher-level abstractions only.

Re-exports

pub use self::adapter::Adapter;

pub use self::adapter::AdapterInfo;

pub use self::adapter::MemoryProperties;

pub use self::adapter::MemoryType;

pub use self::adapter::MemoryTypeId;

pub use self::adapter::PhysicalDevice;

pub use self::adapter::QueuePriority;

pub use self::device::Device;

pub use self::pool::CommandPool;

pub use self::pso::read_spirv;

pub use self::queue::Capability;

pub use self::queue::CommandQueue;

pub use self::queue::Compute;

pub use self::queue::General;

pub use self::queue::Graphics;

pub use self::queue::QueueFamily;

pub use self::queue::QueueGroup;

pub use self::queue::QueueType;

pub use self::queue::Submission;

pub use self::queue::Supports;

pub use self::queue::Transfer;

pub use self::window::AcquireError;

pub use self::window::CompositeAlpha;

pub use self::window::PresentMode;

pub use self::window::Surface;

pub use self::window::SurfaceCapabilities;

pub use self::window::SwapImageIndex;

pub use self::window::Swapchain;

pub use self::window::SwapchainConfig;

Modules

adapter
Physical devices and adapters.

buffer
Memory buffers.

command
Command buffers.

device
Logical device

error
Return values from function calls.

format
Universal format specification. Applicable to textures, views, and vertex buffers.

image
Image related structures.

mapping
Memory mapping

memory
Types to describe the properties of memory allocated for gfx resources.

pass
RenderPass handling.

pool
Command pools

pso
Raw Pipeline State Objects

query
Queries are commands that can be submitted to a command buffer to record statistics or other useful values as the command buffer is running. They are often intended for profiling or other introspection, providing a mechanism for the command buffer to record data about its operation as it is running.

queue
Command queues.

range
Generic range type abstraction that allows ranges to be handled a little more generically.

window
Windowing system interoperability

Macros

spec_const_list
Macro for specifying list of specialization constatns for EntryPoint.

Structs

Features
Features that the device supports. These only include features of the core interface and not API extensions.

Gpu
Represents a combination of a logical device and the hardware queues it provides.

Limits
Resource limits of a particular graphics device.

Enums

IndexType
An enum describing the type of an index value in a slice's index buffer

Primitive
Describes the type of geometric primitives, created from vertex data.

SubmissionError
Marks that an error occurred submitting a command to a command buffer.

Traits

Backend
The Backend trait wraps together all the types needed for a graphics backend. Each backend module, such as OpenGL or Metal, will implement this trait with its own concrete types.

DescriptorPool
A descriptor pool is a collection of memory from which descriptor sets are allocated.

Instance
An instantiated backend.

Type Definitions

AttributeSlot
Slot for an attribute.

ColorSlot
Slot for an active color buffer.

ConstantBufferSlot
Slot for a constant buffer object.

DrawCount
Indirect draw calls count.

IndexCount
Draw number of indices.

InstanceCount
Draw number of instances.

PatchSize
Number of vertices in a patch

ResourceViewSlot
Slot for a shader resource view.

SamplerSlot
Slot for a sampler.

SubmissionResult
Submission result for DX11 backend. Currently mostly unused.

UnorderedViewSlot
Slot for an unordered access object.

VertexCount
Draw vertex count.

VertexOffset
Draw vertex base offset.

WorkGroupCount
Number of work groups.