Module gfx_hal::buffer

Memory buffers.

Buffers interpret memory slices as linear contiguous data array.

They can be used as shader resources, vertex buffers, index buffers or for specifying the action commands for indirect execution.

Structs

Access
Buffer access flags.

IndexBufferView
Index buffer view for bind_index_buffer.

SubRange
A subrange of the buffer.

Usage
Buffer usage flags.

Enums

CreationError
Error creating a buffer.

ViewCreationError
Error creating a buffer view.

Type Definitions

Offset
An offset inside a buffer, in bytes.

State
Buffer state.