Module gfx_hal::command

Command buffers.

A command buffer collects a list of commands to be submitted to the device.

Each command buffer has specific capabilities for graphics, compute or transfer operations, and can be either a primary command buffer or a secondary command buffer.

Operations are always initiated by a primary command buffer, but a primary command buffer can contain calls to secondary command buffers that contain snippets of commands that do specific things, similar to function calls.

All the possible commands are exposed by the CommandBuffer trait.

Structs

BufferCopy
Specifies a source region and a destination region in a buffer for copying. All values are in units of bytes.

BufferImageCopy
Bundles together all the parameters needed to copy a buffer to an image or vice-versa.

ClearDepthStencil
A combination of depth and stencil clear values.

CommandBufferFlags
Option flags for various command buffer settings.

CommandBufferInheritanceInfo

ImageBlit
Parameters for an image blit operation, where a portion of one image is copied into another, possibly with scaling and filtering.

ImageCopy
Bundles together all the parameters needed to copy data from one Image to another.

ImageResolve
Parameters for an image resolve operation, where a multi-sampled image is copied into a single-sampled image.

Enums

AttachmentClear
Attachment clear description for the current subpass.

Level
An enum that indicates whether a command buffer is primary or secondary.

SubpassContents
Specifies how commands for the following renderpasses will be recorded.

Traits

CommandBuffer
A trait that describes all the operations that must be provided by a Backend's command buffer.

Type Definitions

DescriptorSetOffset
Offset for dynamic descriptors.

Unions

ClearColor
A clear color union, which can be either f32, i32, or u32.

ClearValue
A set of clear values for a single attachment.