Module gfx_hal::command::clear

Structs

ClearDepthStencil
A combination of depth and stencil clear values.

Enums

AttachmentClear
Attachment clear description for the current subpass.

Unions

ClearColor
A clear color union, which can be either f32, i32, or u32.

ClearValue
A set of clear values for a single attachment.