Struct gfx_hal::command::SubpassCommandBuffer

pub struct SubpassCommandBuffer<B: Backend, S: Shot, R = <B as Backend>::CommandBuffer>(_, _);

A secondary command buffer recorded entirely within a subpass.

Implementations

impl<B: Backend, S: Shot> SubpassCommandBuffer<B, S>

pub unsafe fn new(raw: B::CommandBuffer) -> Self

Wraps the given CommandBuffer in a SubpassCommandBuffer, starting to record a new subpass.

pub unsafe fn finish(&mut self)

Finish recording commands to the command buffer.

The command pool must be reset to able to re-record commands.

impl<B: Backend> SubpassCommandBuffer<B, OneShot>

pub unsafe fn begin<'a>(
&mut self,
subpass: Subpass<'a, B>,
framebuffer: Option<&'a B::Framebuffer>
)

Begin recording a one-shot sub-pass command buffer.

impl<B: Backend> SubpassCommandBuffer<B, MultiShot>

pub unsafe fn begin<'a>(
&mut self,
allow_pending_resubmit: bool,
subpass: Subpass<'a, B>,
framebuffer: Option<&'a B::Framebuffer>
)

Begin recording a one-shot sub-pass command buffer.