Module gfx_hal::command::structs

Structs

BufferCopy
Specifies a source region and a destination region in a buffer for copying. All values are in units of bytes.

BufferImageCopy
Bundles together all the parameters needed to copy a buffer to an image or vice-versa.

ImageBlit
Parameters for an image blit operation, where a portion of one image is copied into another, possibly with scaling and filtering.

ImageCopy
Bundles together all the parameters needed to copy data from one Image to another.

ImageResolve
Parameters for an image resolve operation, where a multi-sampled image is copied into a single-sampled image.