Module gfx_hal::device

Logical graphics device.

Device

This module exposes the Device trait, which provides methods for creating and managing graphics resources such as buffers, images and memory.

The Adapter and Device types are very similar to the Vulkan concept of "physical devices" vs. "logical devices"; an Adapter is single GPU (or CPU) that implements a backend, a Device is a handle to that physical device that has the requested capabilities and is used to actually do things.

Structs

DeviceLost
Error occurred caused device to be lost.

SurfaceLost
Error occurred caused surface to be lost.

WindowInUse
Native window is already in use by graphics API.

Enums

AllocationError
Possible cause of allocation failure.

BindError
Error binding a resource to memory allocation.

CreationError
Device creation errors during open.

MapError
Error accessing a mapping.

OomOrDeviceLost
Error occurred caused device to be lost or out of memory error.

OutOfMemory
Error allocating memory.

ShaderError
An error from creating a shader module.

WaitFor
Specifies the waiting targets.

Traits

Device
Logical device handle, responsible for creating and managing resources for the physical device it was created from.