Enum gfx_hal::format::SurfaceType

#[repr(u8)]pub enum SurfaceType {
    R4_G4,
    R4_G4_B4_A4,
    B4_G4_R4_A4,
    R5_G6_B5,
    B5_G6_R5,
    R5_G5_B5_A1,
    B5_G5_R5_A1,
    A1_R5_G5_B5,
    R8,
    R8_G8,
    R8_G8_B8,
    B8_G8_R8,
    R8_G8_B8_A8,
    B8_G8_R8_A8,
    A8_B8_G8_R8,
    A2_R10_G10_B10,
    A2_B10_G10_R10,
    R16,
    R16_G16,
    R16_G16_B16,
    R16_G16_B16_A16,
    R32,
    R32_G32,
    R32_G32_B32,
    R32_G32_B32_A32,
    R64,
    R64_G64,
    R64_G64_B64,
    R64_G64_B64_A64,
    B10_G11_R11,
    E5_B9_G9_R9,
    D16,
    X8D24,
    D32,
    S8,
    D16_S8,
    D24_S8,
    D32_S8,
    BC1_RGB,
    BC1_RGBA,
    BC2,
    BC3,
    BC4,
    BC5,
    BC6,
    BC7,
    ETC2_R8_G8_B8,
    ETC2_R8_G8_B8_A1,
    ETC2_R8_G8_B8_A8,
    EAC_R11,
    EAC_R11_G11,
    ASTC_4x4,
    ASTC_5x4,
    ASTC_5x5,
    ASTC_6x5,
    ASTC_6x6,
    ASTC_8x5,
    ASTC_8x6,
    ASTC_8x8,
    ASTC_10x5,
    ASTC_10x6,
    ASTC_10x8,
    ASTC_10x10,
    ASTC_12x10,
    ASTC_12x12,
}

Type of the allocated texture surface. It is supposed to only carry information about the number of bits per each channel. The actual types are up to the views to decide and interpret. The actual components are up to the swizzle to define.

Variants

R4_G4

R4_G4_B4_A4

B4_G4_R4_A4

R5_G6_B5

B5_G6_R5

R5_G5_B5_A1

B5_G5_R5_A1

A1_R5_G5_B5

R8

R8_G8

R8_G8_B8

B8_G8_R8

R8_G8_B8_A8

B8_G8_R8_A8

A8_B8_G8_R8

A2_R10_G10_B10

A2_B10_G10_R10

R16

R16_G16

R16_G16_B16

R16_G16_B16_A16

R32

R32_G32

R32_G32_B32

R32_G32_B32_A32

R64

R64_G64

R64_G64_B64

R64_G64_B64_A64

B10_G11_R11

E5_B9_G9_R9

D16

X8D24

D32

S8

D16_S8

D24_S8

D32_S8

BC1_RGB

BC1_RGBA

BC2

BC3

BC4

BC5

BC6

BC7

ETC2_R8_G8_B8

ETC2_R8_G8_B8_A1

ETC2_R8_G8_B8_A8

EAC_R11

EAC_R11_G11

ASTC_4x4

ASTC_5x4

ASTC_5x5

ASTC_6x5

ASTC_6x6

ASTC_8x5

ASTC_8x6

ASTC_8x8

ASTC_10x5

ASTC_10x6

ASTC_10x8

ASTC_10x10

ASTC_12x10

ASTC_12x12

Implementations

impl SurfaceType

pub fn describe_bits(&self) -> FormatBits

Return the bits for this format.

pub fn desc(&self) -> FormatDesc

Return the format descriptor.