Enum gfx_hal::image::Layout

pub enum Layout {
    General,
    ColorAttachmentOptimal,
    DepthStencilAttachmentOptimal,
    DepthStencilReadOnlyOptimal,
    ShaderReadOnlyOptimal,
    TransferSrcOptimal,
    TransferDstOptimal,
    Undefined,
    Preinitialized,
    Present,
}

Specifies options for how memory for an image is arranged. These are hints to the GPU driver and may or may not have actual performance effects, but describe constraints on how the data may be used that a program must obey. They do not specify how channel values or such are laid out in memory; the actual image data is considered opaque.

Details may be found in the Vulkan spec

Variants

General

General purpose, no restrictions on usage.

ColorAttachmentOptimal

Must only be used as a color attachment in a framebuffer.

DepthStencilAttachmentOptimal

Must only be used as a depth attachment in a framebuffer.

DepthStencilReadOnlyOptimal

Must only be used as a depth attachment in a framebuffer, or as a read-only depth or stencil buffer in a shader.

ShaderReadOnlyOptimal

Must only be used as a read-only image in a shader.

TransferSrcOptimal

Must only be used as the source for a transfer command.

TransferDstOptimal

Must only be used as the destination for a transfer command.

Undefined

No layout, does not support device access. Only valid as a source layout when transforming data to a specific destination layout or initializing data. Does NOT guarentee that the contents of the source buffer are preserved.

Preinitialized

Like Undefined, but does guarentee that the contents of the source buffer are preserved.

Present

The layout that an image must be in to be presented to the display.