Module gfx_hal::pass

Render pass handling.

A render pass represents a collection of

attachments

subpasses

dependencies between the subpasses

and describes how the attachments are used over the course of the subpasses.

Structs

Attachment
An attachment is a description of a resource provided to a render subpass.

AttachmentOps
Attachment operations.

Subpass
A sub-pass borrow of a pass.

SubpassDependency
Expresses a dependency between multiple subpasses.

SubpassDesc
Description of a subpass for render pass creation.

Enums

AttachmentLoadOp
Specifies the operation to be used when reading data from a subpass attachment.

AttachmentStoreOp
Specifies the operation to be used when writing data to a subpass attachment.

Constants

ATTACHMENT_UNUSED
An AttachmentId that can be used instead of providing an attachment.

Type Definitions

AttachmentId
Index of an attachment within a framebuffer/renderpass,

AttachmentLayout
Image layout of an attachment.

AttachmentRef
Reference to an attachment by index and expected image layout.

SubpassId
Index of a subpass.