Module gfx_hal::pso::input_assembler

Input Assembler (IA) stage description. The input assembler collects raw vertex and index data.

Structs

AttributeDesc
Vertex attribute description. Notably, completely separate from resource Descriptors used in DescriptorSets.

Element
A struct element descriptor.

InputAssemblerDesc
All the information needed to create an input assembler.

VertexBufferDesc
Vertex buffer description. Notably, completely separate from resource Descriptors used in DescriptorSets.

Enums

Primitive
Describes the type of geometric primitives, created from vertex data.

VertexInputRate
The rate at which to advance input data to shaders for the given buffer

Type Definitions

BufferIndex
Index of a vertex buffer.

ElemOffset
Offset of an attribute from the start of the buffer, in bytes

ElemStride
Offset between attribute values, in bytes

InstanceRate
Number of instances between each advancement of the vertex buffer.

Location
Shader binding location.

PatchSize
Number of vertices in a patch