Module gfx_hal::pso::output_merger

Output Merger (OM) stage description. The final stage in a pipeline that creates pixel colors from the input shader results, depth/stencil information, etc.

Structs

BlendState
Specifies whether to use blending, and if so, which operations to use for color and alpha channels.

ColorBlendDesc
PSO color target descriptor.

ColorMask
Target output color mask.

DepthStencilDesc
PSO depth-stencil target descriptor.

DepthTest
Depth test state.

Face
Face.

Sided
A generic struct holding the properties of two sides of a polygon.

StencilFace
Complete stencil state for a given side of a face.

StencilTest
Defines a stencil test. Stencil testing is an operation performed to cull fragments; the new fragment is tested against the value held in the stencil buffer, and if the test fails the fragment is discarded.

Enums

BlendOp
Blending operations.

Comparison
A pixel-wise comparison function.

Factor
Defines the possible blending factors. During blending, the source or destination fragment may be multiplied by a factor to produce the final result.

StencilOp
The operation to use for stencil masking.

Type Definitions

StencilValues
Pair of stencil values that could be either baked into a graphics pipeline or provided dynamically.