Module gfx_hal::queue

Command queues.

Queues are the execution paths of the graphical processing units. These process submitted commands buffers.

There are different types of queues, which can only handle associated command buffers. CommandQueue<B, C> has the capability defined by C: graphics, compute and transfer.

Re-exports

pub use self::family::QueueFamily;

pub use self::family::QueueFamilyId;

pub use self::family::QueueGroup;

Modules

family
Queue family and groups.

Structs

Submission
Submission information for a command queue.

Enums

QueueType
The type of the queue, an enum encompassing queue::Capability

Traits

CommandQueue
Abstraction for an internal GPU execution engine.

Type Definitions

QueuePriority
Scheduling hint for devices about the priority of a queue. Values range from 0.0 (low) to 1.0 (high).