Module gfx_hal::queue::capability

Type system encoded queue capabilities.

Enums

Compute
Compute capability, supporting compute and transfer operations.

General
General capability, supporting graphics, compute and transfer operations.

Graphics
Graphics capability, supporting graphics and transfer operations.

GraphicsOrCompute
Graphics or compute capability.

Transfer
Transfer capability, supporting only transfer operations.

Traits

Capability
A Capability is an object that specifies what kind of operations a queue type performs, allowing what types support what queue operations to be described at runtime by the type system.

Supports
A trait that indicates that a particular type of queue supports a particular Capability.

Upper
Encoding the minimal capability to support a combination of other capabilities.