Module gfx_hal::adapter

Physical graphics devices.

The PhysicalDevice trait specifies the API a backend must provide for dealing with and querying a physical device, such as a particular GPU.

An adapter is a struct containing a physical device and metadata for a particular GPU, generally created from an instance of that backend.

Structs

Adapter

Information about a graphics device, supported by the backend.

AdapterInfo

Metadata about a backend adapter.

Gpu

Represents a combination of a logical device and the hardware queues it provides.

MemoryProperties

Types of memory supported by this adapter and available memory.

MemoryType

A description for a single chunk of memory in a heap.

Enums

DeviceType

The type of a physical graphics device

Traits

PhysicalDevice

Represents a physical device (such as a GPU) capable of supporting the given backend.