Trait gfx_hal::pool::CommandPool

pub trait CommandPool<B: Backend>: Debug + Any + Send + Sync {
    unsafe fn reset(&mut self, release_resources: bool);

    unsafe fn free<I>(&mut self, buffers: I)
    where
        I: IntoIterator<Item = B::CommandBuffer>;

    unsafe fn allocate_one(&mut self, level: Level) -> B::CommandBuffer { ... }

    unsafe fn allocate<E>(&mut self, num: usize, level: Level, list: &mut E)
    where
        E: Extend<B::CommandBuffer>,
    { ... }
}

The allocated command buffers are associated with the creating command queue.

Required methods

unsafe fn reset(&mut self, release_resources: bool)

Reset the command pool and the corresponding command buffers.

Synchronization: You may not free the pool if a command buffer is still in use (pool memory still in use)

unsafe fn free<I>(&mut self, buffers: I) where
I: IntoIterator<Item = B::CommandBuffer>,

Free command buffers which are allocated from this pool.

Loading content...
Provided methods

unsafe fn allocate_one(&mut self, level: Level) -> B::CommandBuffer

Allocate a single command buffers from the pool.

unsafe fn allocate<E>(&mut self, num: usize, level: Level, list: &mut E) where
E: Extend<B::CommandBuffer>,

Allocate new command buffers from the pool.

Loading content...