Module gfx_hal::pso::descriptor

Descriptor sets and layouts.

A Descriptor is an object that describes the connection between a resource, such as an Image or Buffer, and a variable in a shader. Descriptors are organized into DescriptorSets, each of which contains multiple descriptors that are bound and unbound to shaders as a single unit. The contents of each descriptor in a set is defined by a DescriptorSetLayout which is in turn built of DescriptorSetLayoutBindings. A DescriptorSet is then allocated from a DescriptorPool using the DescriptorSetLayout, and specific Descriptors are then bound to each binding point in the set using a DescriptorSetWrite and/or DescriptorSetCopy. Each descriptor set may contain descriptors to multiple different sorts of resources, and a shader may use multiple descriptor sets at a time.

Structs

DescriptorPoolCreateFlags

Descriptor pool creation flags.

DescriptorRangeDesc

Set of descriptors of a specific type.

DescriptorSetCopy

Copies a range of descriptors to be bound from one descriptor set to another.

DescriptorSetLayoutBinding

Information about the contents of and in which stages descriptors may be bound to a descriptor set at a certain binding point. Multiple DescriptorSetLayoutBindings are assembled into a DescriptorSetLayout, which is then allocated into a DescriptorSet using a DescriptorPool.

DescriptorSetWrite

Writes the actual descriptors to be bound into a descriptor set.

Enums

AllocationError

An error allocating descriptor sets from a pool.

BufferDescriptorFormat

Format of a buffer.

BufferDescriptorType

Specific type of a buffer.

Descriptor

A handle to a specific shader resource that can be bound for use in a DescriptorSet. Usually provided in a DescriptorSetWrite

DescriptorType

The type of a descriptor.

ImageDescriptorType

Specific type of an image descriptor.

Traits

DescriptorPool

A descriptor pool is a collection of memory from which descriptor sets are allocated.

Type Definitions

DescriptorArrayIndex

DescriptorBinding

DescriptorSetIndex