Module gfx_hal::pso::graphics

Graphics pipeline descriptor.

Structs

BakedStates

Baked-in pipeline states.

BlendDesc

A description of an equation for how to blend transparent, overlapping fragments.

ClearRect

A simple struct describing a rect with integer coordinates.

DepthBias

A depth bias allows changing the produced depth values for fragments slightly but consistently. This permits drawing of multiple polygons in the same plane without Z-fighting, such as when trying to draw shadows on a wall.

GraphicsPipelineDesc

A description of all the settings that can be altered when creating a graphics pipeline.

Multisampling

Rasterizer

Rasterization state.

Rect

A simple struct describing a rect with integer coordinates.

Viewport

A viewport, generally equating to a window on a display.

Enums

FrontFace

The front face winding order of a set of vertices. This is the order of vertexes that define which side of a face is the “front”.

LogicOp

Logic operations used for specifying blend equations.

PolygonMode

Methods for rasterizing polygons, ie, turning the mesh into a raster image.

PrimitiveAssemblerDesc

Primitive Assembler describes how input data are fetched in the pipeline and formed into primitives before being sent into the fragment shader.

Type Definitions

ColorValue

A single RGBA float color.

DepthValue

A single depth value from a depth buffer.

SampleMask

StencilValue

A single value from a stencil buffer.