Trait gfx_hal::queue::family::QueueFamily

pub trait QueueFamily: Debug + Any + Send + Sync {
    fn queue_type(&self) -> QueueType;

    fn max_queues(&self) -> usize;

    fn id(&self) -> QueueFamilyId;
}

General information about a queue family, available upon adapter discovery.

Note that a backend can expose multiple queue families with the same properties.

Can be obtained from an adapter through its queue_families field.

Required methods

fn queue_type(&self) -> QueueType

Returns the type of queues.

fn max_queues(&self) -> usize

Returns maximum number of queues created from this family.

fn id(&self) -> QueueFamilyId

Returns the queue family ID.